定时更新代码

+++ python/webfs/dirs/dirnode.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # --*-- coding:utf-8 --*--
 #目录类
-
+import pdb
 import encrypt
 import jt_global
 import jt_list
@@ -12,13 +12,14 @@ class dirnode(object):
 	#目录的权限
 	__power=666
 	#目录中保存的条目 __childs=0
+	__childs=0
 	#目录不够的时候使用链接下n个块的位置
 	__dirnexts=0
 	#根据名字加密的主键
 	__key=""
 	#初始化目录节点，传入的参数有两个：目录名name,权限控制power	
 	def __init__(self,name,power=666):
-		self.__name=name
+		self.__name=str(name)
 		self.__power=power
 		self.__key=encrypt.jiami(self.__name)
 		self.__childs=jt_list.xlist()
@@ -43,20 +44,38 @@ class dirnode(object):
 
 	#长路径生成目录，如果不存在则一路生成下去	
 	def mkdir2(self,path):
-		names=path.split('/')
-		print names
-		temp=self
-		for item in names:
-			if item!="":
-				res=temp[item]
-				if isinstance(res,dirnode):
-					temp=res
-					continue
-				else:
-					if not temp.mkdir(item):
-						print "wrong"
-					temp=temp[item]
-					continue
+		try:
+			names=path.split('/')
+			temp=self
+			for item in names:
+				if item!="" and temp:
+					res=temp[item]
+					if isinstance(res,dirnode):
+       	                                	temp=res
+       	                                	continue
+       	                         	else:
+       	                                	code=temp.mkdir(item)
+						if code<0:
+       	                                		print "wrong"
+       	                                	res=temp[item]
+						if not res:
+							print '-------------------------'
+							print res
+							print path
+							print item
+							print encrypt.jiami(item)
+							print temp._dirnode__dirnexts[4].ls()
+							print temp._dirnode__dirnexts[4][0].getKey()
+							print temp._dirnode__dirnexts[4][0].getName()
+							pdb.set_trace()
+							print temp['algorithm']
+							print '-------------------------'
+							return False
+						temp=res
+                               	        	continue               
+			return True
+		except Exception,e:
+			return False
 
 	#新增目录
 	def mkdir(self,name):
@@ -65,12 +84,15 @@ class dirnode(object):
 
 		#第一级目录没有满，可以放到里面.
 		if self.__childs.getLength()<jt_global.dir_size:
-			self.__childs.insert(name,temp_dir)
-			return True
+			if self.__childs.insert(name,temp_dir)<0:
+				return -1
+			else:
+				return 0
 
 		#如果第一级目录满了，并且key小于最大值则插入进去后取出最大值放到下一个块中
 		elif temp_dir.getKey()<self.__childs.getMax():
-			self.__childs.insert(name,temp_dir)
+			if self.__childs.insert(name,temp_dir)<0:
+				return -2
 			temp_max=self.__childs.max()
 			self.__childs.deleteByName(temp_max.getName())
 		else:
@@ -80,7 +102,8 @@ class dirnode(object):
 		if temp_max:
 			if self.__dirnexts.getLength()==0:
 				temp_next_dir=dirnext(name,0,0)
-				self.__dirnexts.insert(temp_max.getName(),temp_next_dir)
+				if self.__dirnexts.insert(temp_max.getName(),temp_next_dir)<0:
+					return -3
 
 			target_block_index=self.__dirnexts.bSearch(temp_max.getKey())
 			
@@ -88,72 +111,122 @@ class dirnode(object):
 				if target_block_index['index']<=0 or self.__dirnexts.getLength()==0:
 					target_block_index['index']=0
 
-				temp_max_key=self.__dirnexts[target_block_index['index']].getMaxKey()
-                                temp_min_key=self.__dirnexts[target_block_index['index']].getMinKey()
-				if temp_max.getKey()>temp_max_key and target_block_index['index']<self.__dirnexts.getLength()-1:
-                                	target_block_index['index']+=1
-                                elif temp_max.getKey()<temp_min_key and target_block_index['index']>=1:
-		                	target_block_index['index']-=1
+			temp_max_key=self.__dirnexts[target_block_index['index']].getMaxKey()
+       		        temp_min_key=self.__dirnexts[target_block_index['index']].getMinKey()
+
+			if temp_max.getKey()>temp_max_key and target_block_index['index']<self.__dirnexts.getLength()-1:
+               		   	target_block_index['index']+=1
+               		elif temp_max.getKey()<temp_min_key and target_block_index['index']>=1:
+		           	target_block_index['index']-=1
 
 			if self.__dirnexts[target_block_index['index']].getLength()<jt_global.dir_next_size:
-				self.__dirnexts[target_block_index['index']].insert(temp_max.getName(),temp_max)
+				if self.__dirnexts[target_block_index['index']].insert(temp_max.getName(),temp_max)<0:
+					return -4
 			else:
 				#如何分裂模块
 				#print target_block_index['index']
 				new_block=self.__dirnexts[target_block_index['index']].split(temp_max.getKey())
 				res=self.__dirnexts.deleteByIndex(target_block_index['index'])
 				if new_block['pre'].getLength()<new_block['next'].getLength():
-					new_block['pre'].insert(temp_max.getName(),temp_max)
+					if new_block['pre'].insert(temp_max.getName(),temp_max)<0:
+						return -1
 				else:
-					new_block['next'].insert(temp_max.getName(),temp_max)
+					if new_block['next'].insert(temp_max.getName(),temp_max)<0:
+						return -5
 
-				self.__dirnexts.insert(new_block['pre'].getMaxName(),new_block['pre'])
-				self.__dirnexts.insert(new_block['next'].getMaxName(),new_block['next'])
-		return True
+				if self.__dirnexts.insert(new_block['pre'].getMaxName(),new_block['pre'])<0:
+					return -6
+				if self.__dirnexts.insert(new_block['next'].getMaxName(),new_block['next'])<0:
+					return -7
+		return 0
 
 	#删除目录
-	def rmdir(self,name):
+	def rmdir(self,path):
+		if path=="":                                                     		
+                	print "empty path"
+                	return
+                                                                                 
+                names=path.split('/')
+                if names[0]=="":#绝对路径查找
+                	if names[1]=="":
+                		print "root path"
+                		is_end=True
+                	else:
+                		name=names[1]
+                		if len(names)>2:
+                			is_end=False
+                			left_path=str.join('/',names[2:])
+                		else:
+                			is_end=True
+                                                                                 
+                else:       #相对路径查找
+                	name=names[0]
+                	if len(names)>1:
+                		is_end=False
+                		left_path=str.join('/',names[1:])
+                	else:
+                		is_end=True                                     
+
 		key=encrypt.jiami(name)
-		if key>=self.__childs.getMin() and key<=self.__childs.getMax():
-			res=self.__childs.deleteByName(name)
-			if not res:
-				print "delete fail a"
-		else:
-			res=self.__dirnexts.bSearch(key)
-			if res['index']<=0:
-				res['index']=0
-
-			temp=self.__dirnexts[res['index']]
-			if not isinstance(temp,bool):
-				temp_max=temp.getMaxKey()
-				temp_min=temp.getMinKey()
-				if key>temp_max and res['index']<temp.getLength()-1:
-					res['index']+=1
-					temp=self.__dirnexts[res['index']]
-				elif key<temp_min and res['index']>=1:
-					res['index']-=1
-					temp=self.__dirnexts[res['index']]
-				print temp,res
-				res=temp.getByKey(key)
-				if not isinstance(res,bool):
-					res=temp.deleteByName(res.getName())
-					if not res:
-						print "delete fail b"
+		if key>=self.__childs.getMin() and key<=self.__childs.getMax():                                                            	
+                	res=self.__childs.bSearch(key)
+                	if res['success']:
+                		#print "found:"+str(self.__childs[res['index']].getName())
+                		if not is_end:#如果不是最后一层
+                			self.__childs[res['index']].rmdir(left_path)
 				else:
-					print "delete fail c"
-			else:
-				print "delete fail d"
-		
-			
+					self.__childs.deleteByIndex(res['index'])
+                	else:
+                		#print "childs not found"
+                		return -1
+                else:
+                	res=self.__dirnexts.bSearch(key)
+                	if res['index']<=0:
+                		res['index']=0
+                                                                                                                                           
+                	if self.__dirnexts.getLength()==0:
+                		return -4
+                                                                                                                                           
+                	temp=self.__dirnexts[res['index']]
+                	if not isinstance(temp,bool):
+                		temp_max=temp.getMaxKey()
+                		temp_min=temp.getMinKey()
 
+                		if key>temp_max and res['index']<temp.getLength()-1:
+                			res['index']+=1
+                			temp=self.__dirnexts[res['index']]
+                		elif key<temp_min and res['index']>=1:
+                			res['index']-=1
+                			temp=self.__dirnexts[res['index']]
+                		res_t=temp.getByKey(key)
+                                if not isinstance(res_t,bool):
+                                #	print "found:"+str(temp.getName())
+                			if not is_end:
+                				res_t.rmdir(left_path)
+					else:
+						res_t.deleteByName(name)
+                                else:
+                                #	print "block not found",目前有个问题就是下一层的没有找到，咋办啊
+                			if temp.getLength()>0:
+                				if not is_end:
+							temp[name].rmdir(left_path)
+                				else:
+							temp.deleteByName(name)
+                			else:
+                				print "key="+str(key)+",index="+str(res['index'])
+                				return -2
+                	else:
+                		#print "not exits index="+str(res['index'])+",lenght="+str(self.__dirnexts.getLength())
+                		return -3
+	
 	#获取目录或者文件
 	def __getitem__(self,path):
 		if path=="":
 			print "empty path"
-			return
+			return False
 
 		names=path.split('/')
-		if names[0]=="":#绝对路径查找
+		if names[0]=="":	#绝对路径查找
 			if names[1]=="":
 				print "root path"
 				is_end=True
@@ -171,8 +244,8 @@ class dirnode(object):
 				is_end=False
 				left_path=str.join('/',names[1:])
 			else:
-				is_end=True
-				
+				is_end=True                                     
+	
 		key=encrypt.jiami(name)
 		if key>=self.__childs.getMin() and key<=self.__childs.getMax():
 			res=self.__childs.bSearch(key)
@@ -183,58 +256,86 @@ class dirnode(object):
 				else:
 					return self.__childs[res['index']][left_path]
 			else:
-				#print "childs not found"
-				return -1
+				return False
 		else:
 			res=self.__dirnexts.bSearch(key)
 			if res['index']<=0:
 				res['index']=0
 
 			if self.__dirnexts.getLength()==0:
-				return -4
+				#print key,self.__childs.getMin(),self.__childs.getMax()
+				#print self.getName(),name,path
+				return False
 
 			temp=self.__dirnexts[res['index']]
 			if not isinstance(temp,bool):
 				temp_max=temp.getMaxKey()
 				temp_min=temp.getMinKey()
 
-				if key>temp_max and res['index']<temp.getLength()-1:
+				if key>temp_max and res['index']<self.__dirnexts.getLength()-1:
 					res['index']+=1
 					temp=self.__dirnexts[res['index']]
 				elif key<temp_min and res['index']>=1:
 					res['index']-=1
 					temp=self.__dirnexts[res['index']]
+			
+				if isinstance(temp,bool):
+					return False
 				res_t=temp.getByKey(key)
+
                                 if not isinstance(res_t,bool):
-                                #	print "found:"+str(temp.getName())
 					if is_end:
 						return res_t
 					else:
 						return res_t[left_path]
                                 else:
-                                #	print "block not found",目前有个问题就是下一层的没有找到，咋办啊
 					if temp.getLength()>0:
 						if is_end:
 							return temp[name]
 						else:
-							return temp[name][left_path]
+							temp_node=temp[name]
+							if temp_node:
+								return temp[name][left_path]
+							else:
+								return False
 					else:
 						print "key="+str(key)+",index="+str(res['index'])
-						return -2
+						return False
 			else:
 				#print "not exits index="+str(res['index'])+",lenght="+str(self.__dirnexts.getLength())
-				return -3
+				return False                                                           
 
 
 	#重命名目录
 	def rename(self,old_name,new_name):
-		node=self[old_name]
-		if not isinstance(node,int):
-			old_key=node.getKey()
-			node.setName(new_name)
-			pass
+		if old_name==new_name:
+			print "name equal"
+			return
+		name_list_1=old_name.split('/')
+		name_list_2=new_name.split('/')
+		if len(name_list_1)==len(name_list_2):
+			i=0
+			while i<len(name_list_1)-1:
+				if name_list_1[i]!=name_list_2[i]:
+					print "wrong new name"
+					return
+				i=i+1
 		else:
-			print "rename fail"	
+			print "wrong new name"
+			return
+
+		old_node=self[old_name]
+		if isinstance(old_node,dirnode):
+			self.mkdir2(new_name)
+			curr_node=self[new_name]
+			if isinstance(curr_node,dirnode):
+				curr_node._dirnode__childs=old_node._dirnode__childs
+				curr_node._dirnode__dirnexts=old_node._dirnode__dirnexts
+				self.rmdir(old_name)
+			else:
+				print "rename fail 1"
+		else:
+			print "rename fail 2"
 
 	#显示目录中的内容
 	def ls(self,level=0):
@@ -253,6 +354,22 @@ class dirnode(object):
 		for item in alls:
 			if item.getLength()>0:
 				item.ls(level+1)
+
+	def ls2(self):
+		try:
+			res=[]
+			alls_childs=self.__childs.getAll()
+			for item in alls_childs:
+				res.append(item.getName())
+			alls_nexts=self.__dirnexts.getAll()
+			for item in alls_nexts:
+				if item.getLength():
+					temp=item.getAll()
+					for temp_item in temp:
+						res.append(temp_item.getName())
+			return res
+		except Exception,e:
+			return False
 
 #目录类中定义的下一个块的位置
 class dirnext(object):
